more GridKeyboardNavigation to its own util file
